refactoring to make user input modular, remane profit_calculator.go
