+++ investment_calculator/investment_calculator.go
@@ -9,21 +9,10 @@ const inflationRate = 2.5 // use const keyword to declare a constant
 
 // main entry point of the program, execution starts here
 func main() {
-	var investmentAmount float64 // use var keyword to declare a variable and specify the type when not initializing
-	var years float64
-	expectedReturnRate := 5.5 // use assignment operator to infer type and initialize the variable - kept this for example
+	investmentAmount := getUserInput("Enter the investment amount: ")
+	years := getUserInput("Enter the number of years: ")
+	expectedReturnRate := getUserInput("Enter the expected return rate: ")
 
-	fmt.Print("Enter the investment amount: ") // use fmt.Println to print a message to the console
-	fmt.Scan(&investmentAmount)                // use fmt.Scan to read input from the user
-
-	fmt.Print("Enter the number of years: ")
-	fmt.Scan(&years)
-
-	fmt.Print("Enter the expected return rate: ")
-	fmt.Scan(&expectedReturnRate)
-
-	// futureValue := investmentAmount * math.Pow(1+expectedReturnRate/100, years)
-	// futureRealValue := futureValue / math.Pow(1+inflationRate/100, years)
 	futureValue, futureRealValue := calculateFutureValues(investmentAmount, expectedReturnRate, years)
 
 	fmt.Printf("The future value of the investment is: %.2f\n", futureValue)
@@ -34,6 +23,12 @@ func main() {
 func calculateFutureValues(investmentAmount, expectedReturnRate, years float64) (fv float64, rfv float64) {
 	fv = investmentAmount * math.Pow(1+expectedReturnRate/100, years)
 	rfv = fv / math.Pow(1+inflationRate/100, years)
-	// return fv, rfv
-	return
+	return fv, rfv
+}
+
+func getUserInput(infoText string) float64 {
+	var userInput float64
+	fmt.Print(infoText)
+	fmt.Scan(&userInput)
+	return userInput
 }
